Move route to his own folder
